requirements.txt
livekit-agents
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain-community
requests
python-dotenv